+++ README.md
@@ -8,8 +8,7 @@ Simply open the HTML file `css-wizardry/index.html` in a browser
 
 Hosted here:
 
-Written in Vanilla JS and Handlebars.
-NB: Uses js modules wihtout Webpack or Babel. Will only work locally if run on a local web server. Will also not work in IE or Edge.
+Written in Vanilla JS and Handlebars. Packaged using Webpack.
 
 ### Install dependencies
 
@@ -18,4 +17,4 @@ NB: Uses js modules wihtout Webpack or Babel. Will only work locally if run on a
 ### Run locally
 
 1. `npm start`
-2. Visit localhost:5000
+2. Visit localhost:9000
