+++ package.json
@@ -2,14 +2,21 @@
   "name": "search-app",
   "version": "1.0.0",
   "description": "",
-  "main": "search-app/main.js",
+  "private": true,
   "scripts": {
-    "start": "serve"
+    "build": "webpack --mode=production",
+    "start": "webpack-dev-server"
   },
   "author": "Leo Danielsson",
   "license": "Unlicensed",
   "devDependencies": {
-    "serve": "^11.1.0"
+    "css-loader": "^3.0.0",
+    "handlebars-loader": "^1.7.1",
+    "serve": "^11.1.0",
+    "style-loader": "^0.23.1",
+    "webpack": "^4.35.3",
+    "webpack-cli": "^3.3.5",
+    "webpack-dev-server": "^3.7.2"
   },
   "dependencies": {
     "handlebars": "^4.1.2"
