+++ src/components/result.css
