+++ src/components/result.handlebars
@@ -1,0 +1,4 @@
+<a href={{scryfall_uri}} target='_blank' rel='noopener noreferrer'>
+  <h3>{{name}}</h3>
+</a>
+<img class="Result-image" src={{image_uris.normal}} alt="{{altText}}" />
