+++ src/components/searchHistory.css
