+++ src/components/searchHistoryEntry.handlebars
@@ -1,0 +1,5 @@
+<li data-timestamp={{timestamp}} class="SearchHistory-entry">
+  <div class="SearchHistory-entryName">{{name}}</div>
+  <div>{{formattedTime}}</div>
+  <button data-js="searchHistoryEntryDeleteButton" aria-label="Delete">✕</button>
+</li>
