+++ src/index.css
@@ -28,3 +28,31 @@ header {
   align-items: center;
   justify-content: center;
 }
+
+.SearchApp {
+  font-size: 18px;
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 0 24px;
+}
+
+@media (min-width: 1024px) {
+  .SearchApp {
+    display: grid;
+    grid-gap: 64px;
+    grid-template-columns: 1fr 1fr;
+  }
+}
+
+.SearchApp-search {
+  display: flex;
+  flex-direction: column;
+}
+
+.SearchApp-input {
+  width: 100%;
+  padding: 24px;
+  border: 2px solid lightgray;
+  font-size: 18px;
+  margin-bottom: 8px;
+}
